Rename to Type file for consistent naming and restructure if-else statement
