+++ function/nodejs_14_x/src/s3objectlambda_event.types.ts
